backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.10.4",
    "chardet==3.0.4",
    "deep-translator==1.11.4",
    "fastapi==0.109.2",
    "ffmpeg-python==0.2.0",
    "fpdf==1.7.2",
    "googletrans==3.1.0a0",
    "httpx==0.13.3",
    "librosa==0.10.1",
    "markdown2==2.4.12",
    "numpy==1.26.4",
    "openai-whisper==20231117",
    "pydantic==2.5.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "reportlab==4.1.0",
    "requests==2.31.0",
    "setuptools==69.0.3",
    "soundfile==0.12.1",
    "torch==2.2.1",
    "uvicorn==0.27.1",
    "weasyprint==60.2",
    "yt-dlp==2025.5.22",
]
